+++ test/ast/json_test.dart
@@ -109,7 +109,7 @@ void main() {
   });
 
   test("verify nullable array getter returns a List", () {
-    print(decoded.arrayNodeOrNull<List<String?>>("nestedness")!.data);
+
     expect(
         decoded.arrayNodeOrNull<List<String?>>("nestedness")!.data[0][0], "a");
     expect(