+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Add hasKey method to JsonObject.
+- Add EnumType to AST.
+- Add support for generating dataclasses from Metadata.
+
 ## 0.0.1
 
 - Initial version.